feat: rewrite handler
