+++ packages/code-du-travail-frontend/src/outils/publicodes/context/index.ts
@@ -1,0 +1,1 @@
+export * from "./publicodes";
