+++ packages/code-du-travail-frontend/src/outils/publicodes/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from "./publicodes";
