+++ packages/code-du-travail-frontend/src/outils/publicodes/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./usePublicodes";
