+++ packages/code-du-travail-frontend/src/outils/publicodes/types/indemnite-licenciement.ts
@@ -1,0 +1,3 @@
+export type PublicodesIndemniteLicenciementResult = {
+  value: number;
+};
