+++ packages/code-du-travail-frontend/src/outils/publicodes/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./common";
