+++ packages/code-du-travail-frontend/src/outils/publicodes/types/preavis-retraite.ts
@@ -1,0 +1,14 @@
+export enum PublicodesUnit {
+  DAY = "jour",
+  DAYS = "jours",
+  WEEK = "semaine",
+  WEEKS = "semaines",
+  MONTH = "mois",
+  EUROS = "€",
+}
+
+export type PublicodesPreavisRetraiteResult = {
+  value: number;
+  unit: PublicodesUnit;
+  valueInDays?: number | undefined;
+};
